backend/requirements.txt
fastapi
uvicorn[standard]
apscheduler
httpx
supabase
python-dotenv
